Fix up finding files in non-existent directories
The previous defadvice for `ido-find-file` had the unintended
side-effect that if you're trying to do find-file in a directory that
doesn't exist yet (because you're going to do `M-x make-directory`
afterwards) then since the file wasn't writable (because it was in a
non-existent directory) the defadvice would kick in and try and load it
via Tramp.

Rather than trying to patch it up, I might as well use the version
defined by crux (which is in turn used by Emacs Prelude and thus more
battle-tested).