Get haskell stuff working
A lot of it is just blind guess/trying and relying upon autoloading
from other req-package calls, but everything seems to be loading now.

Originally had tried to use the functionality documented in
https://github.com/edvorg/req-package/issues/18 for more closely
specifying each individual section but couldn't get this to
work (e.g. it worked for the subword aspect but not for shm, etc.).